refactor(code_lines): share the line counter between the size rules
`count_code_lines` was private to `overly_long_function`, but counting
the lines of code in a stretch of source is not particular to a
function body: a rule that caps the length of a file wants the same
count, and wants it to agree. It moves to `src/code_lines.rs` with its
tests.

The counter has an invariant worth its own module docstring -- a
newline inside a token ends the line, which is what makes a multi-line
string count its lines -- so it earns a file rather than a place in
`common`.

Moving it exposed a disagreement in the shipped docs: the rule's
rustdoc did not name doc comments among the lines that are free,
though the counter has always treated them so. It names them now, and
the counter's own tests pin it in both the `///` and `/** */` forms;
before, only an integration fixture reached that behaviour.
