+++ src/rules/overly_long_function.rs
@@ -1,3 +1,4 @@
+use crate::code_lines::count_code_lines;
 use crate::common::DefaultState;
 use crate::measured_fn::measured_fn;
 use crate::rule_index::{Register, rule};
@@ -6,7 +7,6 @@ use clippy_utils::source::snippet_opt;
 use rustc_hir as hir;
 use rustc_hir::def_id::LocalDefId;
 use rustc_hir::intravisit::FnKind;
-use rustc_lexer::{FrontmatterAllowed, TokenKind, tokenize};
 use rustc_lint::{LateContext, LateLintPass, LintStore};
 use rustc_session::{declare_tool_lint, impl_lint_pass};
 use rustc_span::Span;
@@ -18,8 +18,8 @@ declare_tool_lint! {
     /// the body when the count is above `max_lines`.
     ///
     /// A line counts when it holds anything other than whitespace and
-    /// comments, so blank lines, comment-only lines, and the lines a
-    /// block comment spans are free. The braces that open and close
+    /// comments, so blank lines, comment-only lines, doc comments, and
+    /// the lines a block comment spans are free. The braces that open and close
     /// the body are not counted. A function produced by a macro is not
     /// measured. A nested function's lines count towards the function
     /// that contains it, since they sit in its body.
@@ -156,40 +156,6 @@ fn body_interior(source: &str) -> &str {
         .strip_prefix('{')
         .and_then(|rest| rest.strip_suffix('}'))
         .unwrap_or(source)
-}
-
-/// How many lines of `source` hold a token that is neither whitespace
-/// nor a comment.
-fn count_code_lines(source: &str) -> usize {
-    let mut code_lines = 0;
-    let mut line_has_code = false;
-    let mut offset = 0;
-    for token in tokenize(source, FrontmatterAllowed::No) {
-        let text = &source[offset..offset + token.len as usize];
-        offset += token.len as usize;
-        let is_code = !matches!(
-            token.kind,
-            TokenKind::Whitespace | TokenKind::LineComment { .. } | TokenKind::BlockComment { .. },
-        );
-        // Each newline inside the token ends a line. The first ends the
-        // line the token started on, which counts if anything before it
-        // was code; the rest end lines lying wholly inside the token,
-        // which count only for a code token such as a multi-line string.
-        for (newline_index, _) in text.match_indices('\n').enumerate() {
-            if is_code || (newline_index == 0 && line_has_code) {
-                code_lines += 1;
-            }
-        }
-        line_has_code = if text.contains('\n') {
-            is_code
-        } else {
-            line_has_code || is_code
-        };
-    }
-    if line_has_code {
-        code_lines += 1;
-    }
-    code_lines
 }
 
 #[cfg(test)]
